Move rpacket_trackers_to_dataframe function to packet_trackers.py file
