+++ benchmarks/transport_montecarlo_r_packet.py
@@ -6,7 +6,7 @@ from copy import deepcopy
 
 from benchmarks.benchmark_base import BenchmarkBase
 from tardis.base import run_tardis
-from tardis.transport.montecarlo.r_packet import (
+from tardis.transport.montecarlo.packet_trackers import (
     rpacket_trackers_to_dataframe,
 )
